+++ src/preview/preview.ts
@@ -110,6 +110,7 @@ class TextilePreview extends Disposable implements WebviewResourceProvider {
 	private readonly _onScrollEmitter = this._register(new vscode.EventEmitter<LastScrollLocation>());
 	public readonly onScroll = this._onScrollEmitter.event;
 
+	private readonly _disposeCts = this._register(new vscode.CancellationTokenSource());
 	constructor(
 		webview: vscode.WebviewPanel,
 		resource: vscode.Uri,
@@ -202,6 +203,7 @@ class TextilePreview extends Disposable implements WebviewResourceProvider {
 	}
 
 	override dispose() {
+		this._disposeCts.cancel();
 		super.dispose();
 		this._disposed = true;
 		clearTimeout(this.throttleTimer);
@@ -284,7 +286,9 @@ class TextilePreview extends Disposable implements WebviewResourceProvider {
 		try {
 			document = await vscode.workspace.openTextDocument(this._resource);
 		} catch {
-			await this.showFileNotFoundError();
+			if (!this._disposed) {
+				await this.showFileNotFoundError();
+			}
 			return;
 		}
 
@@ -303,7 +307,7 @@ class TextilePreview extends Disposable implements WebviewResourceProvider {
 		const shouldReloadPage = forceUpdate || !this.currentVersion || this.currentVersion.resource.toString() !== pendingVersion.resource.toString() || !this._webviewPanel.visible;
 		this.currentVersion = pendingVersion;
 		const content = await (shouldReloadPage
-			? this._contentProvider.provideTextDocumentContent(document, this, this._previewConfigurations, this.line, this.state)
+			? this._contentProvider.provideTextDocumentContent(document, this, this._previewConfigurations, this.line, this.state, this._disposeCts.token)
 			: this._contentProvider.textileBody(document, this));
 
 		// Another call to `doUpdate` may have happened.