+++ src/test/rename.test.ts
@@ -10,11 +10,12 @@ import { TextileLinkProvider } from '../languageFeatures/documentLinkProvider';
 import { TextileReferencesProvider } from '../languageFeatures/references';
 import { TextileRenameProvider, TextileWorkspaceEdit } from '../languageFeatures/rename';
 import { githubSlugifier } from '../slugify';
+import { noopToken } from '../util/cancellation';
 import { InMemoryDocument } from '../util/inMemoryDocument';
 import { TextileWorkspaceContents } from '../workspaceContents';
 import { createNewTextileEngine } from './engine';
 import { InMemoryWorkspaceTextileDocuments } from './inMemoryWorkspace';
-import { assertRangeEqual, joinLines, noopToken, workspacePath } from './util';
+import { assertRangeEqual, joinLines, workspacePath } from './util';
 
 
 /**
@@ -468,19 +469,19 @@ suite('textile: rename', () => {
 			doc1,
 			doc2
 		]));
-		assertEditsEqual(edit!, {
-			originalUri: workspacePath('images', 'more', 'image.png'),
-			newUri: workspacePath('img', 'test', 'new.png'),
-		}, {
-			uri: uri1, edits: [
-				new vscode.TextEdit(new vscode.Range(0, 1, 0, 23), '/img/test/new.png'),
-				new vscode.TextEdit(new vscode.Range(2, 5, 2, 27), '/img/test/new.png'),
-			]
-		}, {
-			uri: uri2, edits: [
-				new vscode.TextEdit(new vscode.Range(0, 1, 0, 23), '/img/test/new.png'),
-			]
-		});
+		assertEditsEqual(edit!,
+			// Should not have file edits since the files don't exist here
+			{
+				uri: uri1, edits: [
+					new vscode.TextEdit(new vscode.Range(0, 1, 0, 23), '/img/test/new.png'),
+					new vscode.TextEdit(new vscode.Range(2, 5, 2, 27), '/img/test/new.png'),
+				]
+			},
+			{
+				uri: uri2, edits: [
+					new vscode.TextEdit(new vscode.Range(0, 1, 0, 23), '/img/test/new.png'),
+				]
+			});
 	});
 
 	test('Path rename should use .textile extension on extension-less link', async () => {
@@ -560,24 +561,27 @@ suite('textile: rename', () => {
 	test('Path rename should resolve on links without prefix', async () => {
 		const uri1 = workspacePath('sub', 'doc.textile');
 		const doc1 = new InMemoryDocument(uri1, joinLines(
-			`!images/cat.gif(text)!`,
+			`!sub2/doc3.textile(text)!`,
 		));
 
-		const uri2 = workspacePath('doc2.md');
+		const uri2 = workspacePath('doc2.textile');
 		const doc2 = new InMemoryDocument(uri2, joinLines(
-			`!sub/images/cat.gif(text)!`,
+			`!sub/sub2/doc3.textile(text)!`,
 		));
 
-		const edit = await getRenameEdits(doc1, new vscode.Position(0, 10), 'img/cat.gif', new InMemoryWorkspaceTextileDocuments([
-			doc1, doc2,
+		const uri3 = workspacePath('sub', 'sub2', 'doc3.textile');
+		const doc3 = new InMemoryDocument(uri3, joinLines());
+
+		const edit = await getRenameEdits(doc1, new vscode.Position(0, 10), 'sub2/cat.textile', new InMemoryWorkspaceTextileDocuments([
+			doc1, doc2, doc3
 		]));
 		assertEditsEqual(edit!, {
-			originalUri: workspacePath('sub', 'images', 'cat.gif'),
-			newUri: workspacePath('sub', 'img', 'cat.gif'),
+			originalUri: workspacePath('sub', 'sub2', 'doc3.textile'),
+			newUri: workspacePath('sub', 'sub2', 'cat.textile'),
 		}, {
-			uri: uri1, edits: [new vscode.TextEdit(new vscode.Range(0, 1, 0, 15), 'img/cat.gif')]
+			uri: uri1, edits: [new vscode.TextEdit(new vscode.Range(0, 1, 0, 18), 'sub2/cat.textile')]
 		}, {
-			uri: uri2, edits: [new vscode.TextEdit(new vscode.Range(0, 1, 0, 19), 'sub/img/cat.gif')]
+			uri: uri2, edits: [new vscode.TextEdit(new vscode.Range(0, 1, 0, 22), 'sub/sub2/cat.textile')]
 		});
 	});